+++ src/main/java/io/nagurea/smsupsdk/contacts/blacklist/body/Blacklist.java
@@ -1,0 +1,12 @@
+package io.nagurea.smsupsdk.contacts.blacklist.body;
+
+import lombok.Builder;
+import lombok.ToString;
+
+@Builder
+@ToString
+public class Blacklist {
+
+    private final ListOfContacts contacts;
+
+}