+++ src/main/java/io/nagurea/smsupsdk/contacts/blacklist/body/ListOfContacts.java
@@ -1,0 +1,14 @@
+package io.nagurea.smsupsdk.contacts.blacklist.body;
+
+import lombok.Builder;
+import lombok.ToString;
+
+import java.util.List;
+
+@Builder
+@ToString
+public class ListOfContacts {
+
+    private final List<Gsm> gsm;
+
+}